change id in repositories to Long; add delete method in meal controller

+++ src/main/java/com/aws/codestar/projecttemplates/dao/IngredientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngredientRepository extends CrudRepository<Ingredient, Integer> {
+public interface IngredientRepository extends CrudRepository<Ingredient, Long> {
 
 }
