+++ src/main/java/com/aws/codestar/projecttemplates/dao/MealRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MealRepository extends CrudRepository<Meal, Integer> {
+public interface MealRepository extends CrudRepository<Meal, Long> {
 
 }
