+++ src/main/java/com/aws/codestar/projecttemplates/dao/ShoppingListRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShoppingListRepository extends CrudRepository<ShoppingList, Integer> {
+public interface ShoppingListRepository extends CrudRepository<ShoppingList, Long> {
 
 }
